# api/openapi/spec.json
{
  "openapi": "3.0.3",
...
  "info": {
    "title": "TempConnect API",
    "version": "2.0.0",
    "description": "B2B Workforce Marketplace API — VMS, Capacity Exchange, Timesheets, RBAC, Compliance.\n\nAuthentication: Session-cookie (`tc.sid`) + `X-CSRF-Token` header for mutating requests.\n\nGet a CSRF token via `GET /api/csrf` before any POST/PATCH/DELETE.\n\n⚠️ NOTICE: This spec covers only the initial pilot routes (27 of 350+ endpoints). See docs/API_DOCUMENTATION.md for current coverage.",
    "contact": {
      "name": "TempConnect Engineering",
      "email": "tech@tempconnect.de"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    { "url": "http://localhost:3000/api", "description": "Local development" },
    { "url": "https://api.tempconnect.de/api", "description": "Production" }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tc.sid",
        "description": "HTTP-only session cookie set on login"
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "CSRF token obtained from GET /api/csrf"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "VALIDATION" },
          "message": { "type": "string" },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": { "type": "array", "items": { "type": "string" } },
                "message": { "type": "string" }
              }
            }
          }
        },
        "required": ["error"]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["company", "agency", "worker"] },
          "company_name": { "type": "string", "nullable": true },
          "phone": { "type": "string", "nullable": true },
          "postal_code": { "type": "string", "nullable": true },
          "city": { "type": "string", "nullable": true },
          "country": { "type": "string", "default": "DE" },
          "lat": { "type": "number", "nullable": true },
          "lng": { "type": "number", "nullable": true },
          "is_verified": { "type": "boolean" },
          "org_id": { "type": "string", "format": "uuid", "nullable": true },
          "plan": { "type": "string", "enum": ["FREE", "BASIS", "PLUS", "PRO", "ENTERPRISE"] },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": [
    { "sessionCookie": [], "csrfToken": [] }
  ],
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Register a new company or agency account",
        "security": [{ "csrfToken": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["role", "email", "password"],
                "properties": {
                  "role": { "type": "string", "enum": ["company", "agency"] },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 8 },
                  "company_name": { "type": "string" },
                  "phone": { "type": "string" },
                  "postal_code": { "type": "string" },
                  "city": { "type": "string" },
                  "plan": { "type": "string", "enum": ["FREE", "BASIS", "PLUS", "PRO", "ENTERPRISE"], "default": "FREE" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/User" },
                    {
                      "type": "object",
                      "properties": { "verification_sent": { "type": "boolean" } }
                    }
                  ]
                }
              }
            }
          },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "409": { "description": "Email already registered", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login",
        "security": [{ "csrfToken": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Logged in", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "401": { "description": "Invalid credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/auth/forgot-password": {
      "post": {
        "tags": ["auth"],
        "summary": "Request password reset email",
        "security": [{ "csrfToken": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "required": ["email"], "properties": { "email": { "type": "string", "format": "email" } } }
            }
          }
        },
        "responses": {
          "200": { "description": "Email sent (always 200 to prevent user enumeration)" }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": ["auth"],
        "summary": "Set new password using reset token",
        "security": [{ "csrfToken": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token", "password"],
                "properties": {
                  "token": { "type": "string" },
                  "password": { "type": "string", "minLength": 8 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Password reset" },
          "400": { "description": "Token expired or password too short" }
        }
      }
    },
...
  }
}